Update as of 10/21/2013 at 1:42 am
So i think i got everything stored in a map the way we wanted.
Made a change with the categories. Did the creating category id part. Check it out and let me know if you like it.

Also did a small test in the main. Has you name in the comments.
Check it out and play around with it.

Tried to comment every line of code. Most was just built on what we did.